+++ crypto-common/CHANGELOG.md
@@ -8,12 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.2.0 (UNRELEASED)
 ### Added
 - Sealed `BlockSizes` trait implemented for types from `U1` to `U255`
+- `Generate` trait ([#2096])
 
 ### Changed
 - `BlockUser::BlockSize` is now bounded by the `BlockSizes` trait
 - Edition changed to 2024 and MSRV bumped to 1.85 ([#1759])
 
+### Removed
+- `generate_*` and `try_generate_*` methods from KeyInit and KeyIvInit traits.
+  See the newly added Generate trait for replacement ([#2096])
+
 [#1759]: https://github.com/RustCrypto/traits/pull/1759
+[#2096]: https://github.com/RustCrypto/traits/pull/2096
 
 ## 0.1.7 (2025-11-12)
 ### Changed